rework functionality a bit - the language-keys will be on div elements with class="key" and data-key="number" number here corresponds to the cue id in the audio's vtt-file of type captions.
this is used to display which language key (number of paragraphs) are currently being played in the audio, and
to sync the translation. These id's will match between the languages.

the other tracks, with kind = subtitles are intended for syncing other language's text and displaying next to the language currently played. The id's will always match in the vtt-files, they are for language keys.

p - elements are used just for paragraphs. No sync here.

for norwegian, more detailed sync is added.
here the span elements with timecode in data-begin and data-end attributes are used to signal what time of the audio of the current time the text passages belong.

So there is a two-level sync. Language key, and sub-sentence level.